+++ src/utils/theme.ts
@@ -1,6 +1,6 @@
 'use client';
 
-import { useLocalStorage } from './localstorage';
+import { useLocalStorage } from '@/hooks/use-local-storage';
 
 export function useTheme() {
   return useLocalStorage('theme', typeof window !== 'undefined' && window.matchMedia('(prefers-color-scheme: dark)').matches ? 'dark' : 'light');
